+++ ftests.py
@@ -1,0 +1,6 @@
+import unittest
+from functional_tests.FYLM141111001 import FunctionalTests
+
+
+if __name__ == '__main__':
+    unittest.main()